Create ExpandableLayoutDelegate
