+++ sample/src/main/res/layout/activity_example_search.xml
@@ -20,7 +20,7 @@
         android:layout_alignParentTop="true"
         android:layout_toLeftOf="@id/expandButton" />
 
-    <com.github.aakira.expandablelayout.ExpandableRelativeLayout
+    <com.github.aakira.expandablelayout.ExpandableFrameLayout
         android:id="@+id/expandableLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -35,5 +35,5 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:scrollbars="vertical" />
-    </com.github.aakira.expandablelayout.ExpandableRelativeLayout>
+    </com.github.aakira.expandablelayout.ExpandableFrameLayout>
 </RelativeLayout>
